+++ xdsl/transforms/apply_pdl_interp.py
@@ -1,12 +1,49 @@
 import os
 from dataclasses import dataclass
+from typing import IO
 
 from xdsl.context import Context
 from xdsl.dialects import builtin, pdl_interp
-from xdsl.interpreters.pdl_interp import PDLInterpRewritePattern
+from xdsl.dialects.builtin import ModuleOp
+from xdsl.interpreter import Interpreter
+from xdsl.interpreters.pdl_interp import PDLInterpFunctions
+from xdsl.ir import Operation
 from xdsl.parser import Parser
 from xdsl.passes import ModulePass
-from xdsl.pattern_rewriter import PatternRewriteWalker
+from xdsl.pattern_rewriter import PatternRewriter, PatternRewriteWalker, RewritePattern
+
+
+@dataclass
+class PDLInterpRewritePattern(RewritePattern):
+    """
+    A rewrite pattern that uses the pdl_interp dialect for matching and rewriting operations.
+    """
+
+    ctx: Context
+    interpreter: Interpreter
+    functions: PDLInterpFunctions
+    matcher: pdl_interp.FuncOp
+
+    def __init__(
+        self, matcher: pdl_interp.FuncOp, ctx: Context, file: IO[str] | None = None
+    ):
+        # Create interpreter and register implementations
+        self.ctx = ctx
+        self.functions = PDLInterpFunctions(ctx)
+        module = matcher.parent_op()
+        assert isinstance(module, ModuleOp)
+        self.interpreter = Interpreter(module=module, file=file)
+        self.interpreter.register_implementations(self.functions)
+        if matcher.sym_name.data != "matcher":
+            raise ValueError("Matcher function name must be 'matcher'")
+        self.matcher = matcher
+
+    def match_and_rewrite(self, xdsl_op: Operation, rewriter: PatternRewriter) -> None:
+        # Setup the rewriter
+        self.functions.rewriter = rewriter
+
+        # Call the matcher function on the operation
+        self.interpreter.call_op(self.matcher, (xdsl_op,))
 
 
 @dataclass(frozen=True)
